+++ PinnyNotes.Core/Enums/NewInstanceBehaviour.cs
@@ -1,0 +1,7 @@
+﻿namespace PinnyNotes.Core.Enums;
+
+public enum NewInstanceBehaviour
+{
+    CreateNewNote = 0,
+    ShowManagementWindow = 1
+}