+++ PinnyNotes.WpfUi/Models/ApplicationSettingsModel.cs
@@ -1,8 +1,12 @@
-﻿namespace PinnyNotes.WpfUi.Models;
+﻿using PinnyNotes.Core.Enums;
+
+namespace PinnyNotes.WpfUi.Models;
 
 public class ApplicationSettingsModel : BaseModel
 {
     // General
+    public StartupBehaviour StartupBehaviour { get; set => SetProperty(ref field, value); }
+    public NewInstanceBehaviour NewInstanceBehaviour { get; set => SetProperty(ref field, value); }
     public bool ShowNotifyIcon { get; set => SetProperty(ref field, value); }
     public bool CheckForUpdates { get; set => SetProperty(ref field, value); }
 }