+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Week Number action to Date Time tool.
 - Fluent theme, other windows (settings, management, etc) and menus now match users Windows theme.
 - Pinned by default setting.
+- Startup and new instance behaviour settings.
 - 
 
 ### Changed